refactor: view를 web과 console로 분리
