+++ src/view/web/LotteriesView.js
