+++ src/view/web/PurchaseView.js
